Move chilipiper to its own hook
