+++ website/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useHubSpot } from './hubSpot'
 export { useCarousel } from './carousel'
 export { useWindowWidth } from './windowWidth'
+export { useChiliPiper } from './chiliPiper'
